refactor: 셀렉트 버튼 외부 요소 클릭을 감지하는 로직 커스텀 훅으로 분리
